sign_tx: update for stax
